refactor(agent-runtime): dedupe OTel test doubles into testing/
RecordingSpan + RecordingTracerProvider were copy-pasted (~120 LOC)
across agent-runner.test.ts and llm-client.test.ts. Extracted the
richer variant (parent tracking, exception recording) to
testing/recording-tracer.ts, exported via testing/index.
