+++ packages/agent-runtime/src/testing/index.ts
@@ -1,2 +1,3 @@
 export { InMemoryAgentEventLog } from './in-memory-agent-event-log';
 export { NoopLlmClient } from './noop-llm-client';
+export { RecordingSpan, RecordingTracerProvider } from './recording-tracer';
